Draft implementation for the opcode decode method
The idea is to decode opcodes to a tuple where the
1st element is the opcode to be mapped to a dict.

Depending on the opcode, the following elements 
will be it's parameters relevant to change states.

By doing this I could consider each instruction a
code unit with the proper parameters avoiding to
run the state transitions inside tons of "if"s.